[nxp noup] platform: Added tfm support for board: rd_rw612_bga
added rd_rw612_bga board support in downstream
fork of tfm branch of nxp.

Few more details are as following:
 - SDK 2.16.000 drivers updated
 - psa-crypto-driver patch from develop: sdk_api_2.x that has the
support of threading apps with mbedtls 3.6.0
 - Support using ELS die internal key as ROT for HUK.
 - Add option to use ELS based key for initial attestation.
 - For rollback protection in ITS, the approach is to use two additional nv
counters. Definition of IDs of nv_counters as well as checking of usage
permissions for those counters is done in TF-M core code and there is no
way for a platform to extend this easily.
 - Also, as the nv counters are implemented as part of the platform service
and get used from the ITS service, a dependency from ITS on the platform
service is required (in a similar fashion as it is required for the
rollback protection of PS).
 - As this requires platform specific changes to TF-M core code, this is
included here as a patch.
 - A new partition is added for RW61x, which can provide apis
to do FW loading service.

Signed-off-by: Waqar Tahir <waqar.tahir@nxp.com>
Signed-off-by: Stefan Krug <stefan.krug@nxp.com>
Signed-off-by: Philipp Wallisch <philipp.wallisch@nxp.com>