[IOTDL-1729] Fix N10 RAM size and flash rule configuration
The flash rule configuration had an off-by-one error, causing hardfaults
when NS code tries to access the last 32KB of the NS region.

The RAM was wrongly configured as it didn't account for the 16KB of
reserved memory for low power & ELS PKC, overflowing the available
416KB (in ECC config) by 16KB.
This commit shrinks both NS & S RAM size by 8KB to 200KB each.

Signed-off-by: Philipp Wallisch <philipp.wallisch@nxp.com>
Signed-off-by: Waqar Tahir <waqar.tahir@nxp.com>