Crypto: Update mbed TLS includes that were missed in the first upgrade
When upgrading from Mbed TLS 3.6.2 to 3.6.3, a cherry-pick was created
from the patch on main, but for some reason a few headers for the
interface/include/mbedtls directory were not upgraded correctly. This
would lead to issue in integration such as SDKs or the Zephyr one, if
both projects are present will lead to header mismatches with code.

Signed-off-by: Antonio de Angelis <antonio.deangelis@arm.com>
Change-Id: I3ab88ddd57b82d3bf5b093955274ad9cdd4e90fc