[MCUX-83272] Add files from upstream to common/Native_Driver
In mcux sdk build, these drivers,components and utilities
are directly picked from mcux-sdk repo.However in upstream,
for standalone build a copy of these is maintained in
nxp/common/Native_Driver.

This commit adds the native drivers which are needed for
MCIMX932 that are already present in tfm upstream.

Signed-off-by: Ruchika Gupta <ruchika.gupta_1@nxp.com>